requirements.txt
mcp>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
gitpython>=3.1.0
pydantic>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
transformers>=4.0.0
textblob>=0.17.0
spacy>=3.0.0
